Move views to templates, add readme for creating views from templates
