+++ README.md
@@ -1,1 +1,14 @@
-This is the `README.md` for the whole monorepo.
+# Micro-Front End Portal
+
+## Getting Started
+Load dependencies and start up the application catalogue server and build the host view.
+```
+yarn && yarn serve
+```
+
+## Create an Application
+
+```
+yarn modular add card-view
+yarn build card-view
+```
