--- packages/card-view/package.json
@@ -1,16 +1,0 @@
-{
-  "name": "card-view",
-  "private": false,
-  "modular": {
-    "type": "esm-view"
-  },
-  "dependencies": {
-    "@jpmorganchase/uitk-core": "^0.6.0",
-    "@jpmorganchase/uitk-grid": "^0.3.0",
-    "@jpmorganchase/uitk-lab": "^0.6.0",
-    "@jpmorganchase/uitk-theme": "^0.5.0",
-    "classnames": "^2.3.2"
-  },
-  "version": "1.0.0",
-  "devDependencies": {}
-}
