+++ packages/templates/card-view/src/react-app-env.d.ts
