+++ packages/templates/card-view/src/util/md5.js
