+++ packages/templates/card-view/src/util/person.ts
