+++ packages/templates/grid-view/src/components/avatar-size-picker.tsx
