+++ packages/templates/grid-view/src/components/person-grid.tsx
