+++ packages/templates/grid-view/src/util/person.ts
