--- packages/view1/package.json
@@ -1,8 +1,0 @@
-{
-  "name": "view1",
-  "private": false,
-  "modular": {
-    "type": "esm-view"
-  },
-  "version": "1.0.0"
-}
