--- packages/view1/src/__tests__/EsmView.test.tsx
@@ -1,9 +1,0 @@
-import * as React from 'react';
-import { render, screen } from '@testing-library/react';
-import EsmView from '..';
-
-test('renders learn react link', () => {
-  render(<EsmView />);
-  const linkElement = screen.getByText(/learn react/i);
-  expect(linkElement).toBeInTheDocument();
-});
