--- packages/view1/src/react-app-env.d.ts
@@ -1,1 +1,0 @@
-/// <reference types="modular-scripts/react-app-env" />
