--- packages/view2/package.json
@@ -1,8 +1,0 @@
-{
-  "name": "view2",
-  "private": false,
-  "modular": {
-    "type": "esm-view"
-  },
-  "version": "1.0.0"
-}
