--- packages/view2/src/index.tsx
@@ -1,6 +1,0 @@
-import * as React from 'react';
-import './index.css';
-
-export default function EsmView(): JSX.Element {
-  return <div>Hi, I'm view 2</div>;
-}
